Add icontains lookup
The problem was that the `as_sql` was not considering that the lhs query
could potentially have params to be evaluated for each of the three
different possibilities. This PR fixes it by introducing a helper method
responsible for the lhs_params extension.
Furthermore, the front end uses icontains when filtering by column,
meaning that IPCharField needs to implement it which is also done in
this PR.